requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
diffusers==0.21.4
transformers==4.35.0
accelerate==0.24.1
pillow==10.1.0
python-multipart==0.0.6
google-generativeai
google-generativeai

